+++ lib/Screens/List_Screen.dart
@@ -1,14 +1,15 @@
+// ignore_for_file: use_build_context_synchronously, file_names
+
+import 'dart:async';
 import 'dart:math';
 
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:randomizer/Database/Models/Items.dart';
 import 'package:randomizer/Database/Models/Listas.dart';
 import 'package:randomizer/Database/RandomizerDB.dart';
 import 'package:randomizer/templates/Molecules/Button.dart';
 import 'package:randomizer/templates/Molecules/Divisor.dart';
-import 'package:randomizer/templates/Molecules/Empty_List.dart';
 import 'package:randomizer/templates/Molecules/Inputs/Input_Gn_T2.dart';
 import 'package:randomizer/templates/Molecules/List_Item.dart';
 
@@ -50,8 +51,48 @@ class _ListScreenState extends State<ListScreen> {
   }
 
   void onPressedDelete(int id) async {
-    await RandomizerDB.deleteItem(id);
-    getInfoItems();
+    int index = items.indexWhere((element) => element.id == id);
+
+    await showDialog(
+      context: context,
+      builder: (context) => AlertDialog(
+        title: Text(
+          "Alerta",
+          textAlign: TextAlign.center,
+          style: TextStyle(
+            color: Theme.of(context).colorScheme.error,
+            fontSize: Theme.of(context).textTheme.headlineLarge!.fontSize,
+            fontFamily: Theme.of(context).textTheme.headlineLarge!.fontFamily,
+          ),
+        ),
+        content: Text(
+            "Está a punto de borrar la opción ${items[index].title} de la lista ${list.title}. "
+            "¿Está seguro de eliminar la opción permanentemente?"),
+        actions: [
+          //Cancelar
+          TextButton(
+            onPressed: () {
+              Navigator.pop(context);
+            },
+            child: const Text("Cancelar"),
+          ),
+          //Cancelar
+          TextButton(
+            onPressed: () async {
+              Navigator.pop(context);
+              await RandomizerDB.deleteItem(id);
+              getInfoItems();
+            },
+            child: Text(
+              "Borrar",
+              style: TextStyle(
+                color: Theme.of(context).colorScheme.error,
+              ),
+            ),
+          ),
+        ],
+      ),
+    );
   }
 
   void addItem() async {
@@ -237,7 +278,16 @@ class _ListScreenState extends State<ListScreen> {
 
           //Accept
           TextButton(
-            onPressed: () {
+            onPressed: () async {
+              if (opcionController.text.isNotEmpty) {
+                Listas auxList = Listas(
+                  id: list.id,
+                  title: opcionController.text,
+                  description: desOpcionController.text,
+                );
+                await RandomizerDB.updateLista(auxList);
+                getInfoList();
+              }
               Navigator.pop(context);
             },
             child: Text(
@@ -250,20 +300,61 @@ class _ListScreenState extends State<ListScreen> {
         ],
       ),
     );
-
-    if (opcionController.text.isNotEmpty) {
-      Listas auxList = Listas(
-        id: list.id,
-        title: opcionController.text,
-        description: desOpcionController.text,
-      );
-      await RandomizerDB.updateLista(auxList);
-      getInfoList();
-    }
     setState(() {
       opcionController.clear();
       desOpcionController.clear();
     });
+  }
+
+  int chooseOption() {
+    double repeat = Random().nextDouble();
+
+    if (repeat < Random().nextDouble()) {
+      Timer(const Duration(seconds: 5), () {});
+      return chooseOption();
+    } else {
+      return Random().nextInt(items.length);
+    }
+  }
+
+  void showDecision() {
+    int option = chooseOption();
+
+    showDialog(
+      context: context,
+      builder: (context) => AlertDialog(
+        title: Text(
+          "Decisión",
+          style: Theme.of(context).textTheme.headlineLarge,
+        ),
+        content: Wrap(
+          children: [
+            Text(
+                "El sistema de Randomizer ha determinado que usted debería elegir la opción ${items[option].title}. ¿Está de acuerdo con la decisión de Randomizer?"),
+          ],
+        ),
+        actions: [
+          TextButton(
+            onPressed: () {
+              Navigator.pop(context);
+              showDecision();
+            },
+            child: const Text("Repetir"),
+          ),
+          TextButton(
+            onPressed: () {
+              Navigator.pop(context);
+            },
+            child: const Text("Aceptar"),
+          ),
+        ],
+      ),
+    );
+  }
+
+  @override
+  void dispose() {
+    super.dispose();
   }
 
   @override
@@ -318,12 +409,18 @@ class _ListScreenState extends State<ListScreen> {
               },
             ),
           ),
+          const SizedBox(
+            height: 20,
+          ),
           ButtonDashed(
-            onTap: null,
+            onTap: showDecision,
             hint: "Random!",
             enabled: items.isNotEmpty,
             background: Theme.of(context).colorScheme.primary,
             colorText: Colors.white,
+          ),
+          const SizedBox(
+            height: 20,
           ),
         ],
       ),
